move functions in train to better files
