+++ graph_utils.py
@@ -1,5 +1,6 @@
 import jax.numpy as jnp
 from typing import Tuple
+import jraph
 
 # from https://chatgpt.com/share/6726811b-96a0-800e-af41-684b211f59b6
 # TODO: support periodic boundary conditions. the graphs made here are NOT periodic (we'd need to take in the lattice paramters for periodic support)
@@ -25,3 +26,18 @@ def radius_graph(positions: jnp.ndarray, radius: float) -> Tuple[jnp.ndarray, jn
     senders, receivers = jnp.where(mask)
 
     return senders, receivers
+
+def prepare_single_graph(pos: jnp.ndarray, radius: float) -> jraph.GraphsTuple:
+    senders, receivers = radius_graph(pos, radius) # make the radius really big so all nodes are connected (just for testing rn. can reduce to 1.1 layer)
+
+    return jraph.batch([
+        jraph.GraphsTuple(
+            nodes=pos,
+            edges=None,
+            globals=None,
+            senders=senders,
+            receivers=receivers,
+            n_node=jnp.array([len(pos)]),
+            n_edge=jnp.array([len(senders)]),
+        )
+    ])
